+++ trunk/MyGUIEngine/include/MyGUI_ItemBox.h
@@ -42,7 +42,7 @@ namespace MyGUI
 		// манипуляции айтемами
 
 		//! Get number of items
-		size_t getItemCount() { return (size_t)mCountItems; }
+		size_t getItemCount() { return mCountItems; }
 
 		//! Insert an item into a array at a specified position
 		void insertItemAt(size_t _index, Any _data = Any::Null);
@@ -210,7 +210,7 @@ namespace MyGUI
 		virtual void updateDropItems();
 		virtual void updateDropItemsState(const DropWidgetState & _state);
 
-		// Обновляет данные о айтемах, при изменении размеров 
+		// Обновляет данные о айтемах, при изменении размеров
 		void updateMetrics();
 
 		// обновляет скролл, по текущим метрикам
@@ -259,7 +259,7 @@ namespace MyGUI
 		// колличество линий
 		int mCountLines;
 		// колличество айтемов всего
-		int mCountItems;
+		size_t mCountItems;
 		// максимальное колличество видимых линий
 		int mCountLineVisible;
